+++ addons/mysterious_egg/static/src/builder_overlay_plugin/builder_overlay_plugin.js
