+++ addons/mysterious_egg/static/src/builder_tab/block_tab.xml
