+++ addons/mysterious_egg/static/src/builder_tab/customize_tab.xml
