+++ addons/mysterious_egg/static/src/snippets_menu.js
