+++ addons/mysterious_egg/static/src/website_builder.js
@@ -4,7 +4,7 @@ import { useService } from "@web/core/utils/hooks";
 import { Editor } from "@html_editor/editor";
 import { MAIN_PLUGINS } from "@html_editor/plugin_sets";
 import { LazyComponent } from "@web/core/assets";
-import { BuilderOverlayPlugin } from "@mysterious_egg/editor/builder_overlay/builder_overlay_plugin";
+import { BuilderOverlayPlugin } from "@mysterious_egg/builder_overlay_plugin/builder_overlay_plugin";
 import { WebsiteSystrayItem } from "./website_systray_item";
 
 export const unslugHtmlDataObject = (repr) => {
@@ -28,12 +28,12 @@ class WebsiteBuilder extends Component {
         this.orm = useService("orm");
         this.websiteContent = useRef("iframe");
 
-        this.editor = new Editor(
-            {
-                Plugins: [...MAIN_PLUGINS, ...BUILDER_PLUGIN],
-            },
-            this.env.services
-        );
+        // this.editor = new Editor(
+        //     {
+        //         Plugins: [...MAIN_PLUGINS, ...BUILDER_PLUGIN],
+        //     },
+        //     this.env.services
+        // );
 
         onWillStart(async () => {
             const slugCurrentWebsite = await this.orm.call("website", "get_current_website");
@@ -63,7 +63,7 @@ class WebsiteBuilder extends Component {
     }
 
     onWebsiteLoaded() {
-        this.editor.attachTo(this.websiteContent.el.contentDocument.body);
+        // this.editor.attachTo(this.websiteContent.el.contentDocument.body);
     }
 }
